+++ crates/bsk-protocol/src/bin/dump-schema.rs
@@ -114,7 +114,8 @@ fn main() {
     dump!(RequestHelpResult, "tool_request_help_result");
 
     dump!(TraceV2, "trace_v2");
-    dump!(TraceV3, "trace");
+    dump!(TraceV3, "trace_v3");
+    dump!(RecordedTrace, "trace");
     dump!(StepV3, "trace_step");
     dump!(RecordStartParams, "tool_record_start_params");
     dump!(RecordStartResult, "tool_record_start_result");
